Move Adapter to other package
