+++ app/src/main/java/com/htoyama/likit/ui/search/assist/Header.java
@@ -7,15 +7,15 @@ import com.htoyama.likit.R;
 public enum Header {
 
   TAG {
-    @Override String displayingText(Context context) {
+    @Override public String displayingText(Context context) {
       return context.getString(R.string.search_assist_header_tag);
     }
   },
   USER {
-    @Override String displayingText(Context context) {
+    @Override public String displayingText(Context context) {
       return context.getString(R.string.search_assist_header_user);
     }
   };
 
-  abstract String displayingText(Context context);
+  public abstract String displayingText(Context context);
 }
