+++ app/src/main/java/com/htoyama/likit/ui/search/assist/listitem/AssistAdapter.java
@@ -1,4 +1,4 @@
-package com.htoyama.likit.ui.search.assist;
+package com.htoyama.likit.ui.search.assist.listitem;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.ViewGroup;
@@ -6,6 +6,7 @@ import android.view.ViewGroup;
 import com.hannesdorfmann.adapterdelegates2.AdapterDelegatesManager;
 import com.htoyama.likit.domain.tag.Tag;
 import com.htoyama.likit.domain.user.User;
+import com.htoyama.likit.ui.search.assist.Assist;
 
 public class AssistAdapter extends RecyclerView.Adapter {
